+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Features
 
+- Adds GraphQL integration ([#5299](https://github.com/getsentry/sentry-react-native/pull/5299))
 - Add new _experimental_ Canvas Capture Strategy for Session Replay ([#5301](https://github.com/getsentry/sentry-react-native/pull/5301))
   - A new screenshot capture strategy that uses Android's Canvas API for more accurate and reliable text and image masking
   - Any .drawText() or .drawBitmap() calls are replaced by rectangles, ensuring no text or images are present in the resulting output
@@ -26,7 +27,6 @@
       ],
     });
     ```
-
 
 ### Dependencies